OCR_API_project/requirements.txt
asgiref==3.5.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
django-filter==21.1
djangorestframework==3.13.1
fuzzywuzzy==0.18.0
google-api-core==2.7.2
google-auth==2.6.5
google-cloud-vision==2.7.2
googleapis-common-protos==1.56.0
grpcio==1.44.0
grpcio-status==1.44.0
idna==3.3
importlib-metadata==4.11.3
inflection==0.5.1
itypes==1.2.0
jarowinkler==1.0.2
Jinja2==3.1.1
Levenshtein==0.18.1
Markdown==3.3.6
MarkupSafe==2.1.1
numpy==1.22.3
openapi-codec==1.3.2
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
pkg_resources==0.0.0
proto-plus==1.20.3
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
rapidfuzz==2.0.9
requests==2.27.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
zipp==3.8.0
